requirements.txt
PySide6>=6.6.0
pandas>=2.0.0
PyInstaller>=6.0.0
Pillow>=11.0.0